requirements.txt
# Mortgage Calculator Telegram Bot - Dependencies
python-telegram-bot==20.3
python-dotenv==1.0.0
